Refactored Usage tests via TestWiresUsageMixin.
