+++ tests/test_instance.py
@@ -20,94 +20,6 @@ from . import helpers
 
 
 
-class TestWiresUtilization(helpers.CallTrackerAssertMixin, unittest.TestCase):
-
-    """
-    Utilization tests for the Wires singleton.
-    """
-
-    def setUp(self):
-
-        self.w = Wiring()
-
-
-    def test_wire_call(self):
-        """
-        Wires the `this` call and calls it.
-        Checks the wired callable was called once with the correct arguments.
-        """
-        tracker = helpers.CallTracker()
-
-        self.w.wire.this.calls_to(tracker)
-        self.w.this()
-
-        self.assertEqual(tracker.call_count, 1, 'call count mismatch')
-        self.assertEqual(tracker.call_args, [
-            ((), {},),
-        ], 'call argument mismatch')
-
-
-    def test_double_wire_call(self):
-        """
-        Wires the `this` call twice and calls it.
-        Checks the wired callable was called twice with the correct arguments.
-        """
-        tracker = helpers.CallTracker()
-
-        self.w.wire.this.calls_to(tracker)
-        self.w.wire.this.calls_to(tracker)
-        self.w.this()
-
-        self.assertEqual(tracker.call_count, 2, 'call count mismatch')
-        self.assertEqual(tracker.call_args, [
-            ((), {},),
-            ((), {},),
-        ], 'call argument mismatch')
-
-
-    def test_multi_wired_call(self):
-        """
-        Wires the `this` call to multiple callables and calls it.
-        Checks that each of the wired callables was called once with the correct
-        arguments.
-        """
-        num_wirings = 10
-
-        trackers = [helpers.CallTracker() for _ in range(num_wirings)]
-
-        for tracker in trackers:
-            self.w.wire.this.calls_to(tracker)
-
-        self.w.this()
-
-        for tracker in trackers:
-            self.assertEqual(tracker.call_count, 1, 'call count mismatch')
-            self.assertEqual(tracker.call_args, [
-                ((), {},),
-            ], 'call argument mismatch')
-
-
-    def test_wire_call_unwire_call(self):
-        """
-        Wires the `this` call to a callable and calls it.
-        Checks the wired callable was called once with the correct arguments.
-        Unwires the `this` callable and calls it again.
-        Checks the wired callable wasn't called again.
-        """
-        tracker = helpers.CallTracker()
-
-        self.w.wire.this.calls_to(tracker)
-        self.w.this()
-
-        self.assert_single_call_no_args(tracker)
-
-        self.w.unwire.this.calls_to(tracker)
-        self.w.this()
-
-        self.assert_single_call_no_args(tracker)
-
-
-
 class TestWiresCalleeFailures(unittest.TestCase):
 
     """
