+++ tests/test_usage_singleton.py
@@ -1,0 +1,26 @@
+# ----------------------------------------------------------------------------
+# Python Wires Tests
+# ----------------------------------------------------------------------------
+# Copyright (c) Tiago Montes.
+# See LICENSE for deatils.
+# ----------------------------------------------------------------------------
+
+
+from __future__ import absolute_import
+
+import unittest
+
+from . import mixin_test_usage, mixin_test_singleton
+
+
+
+class TestWiresUtilization(mixin_test_singleton.TestWiresSingletonMixin,
+                           mixin_test_usage.TestWiresUsageMixin,
+                           unittest.TestCase):
+
+    """
+    Utilization tests for the Wires singleton.
+    """
+
+
+# ----------------------------------------------------------------------------
